requirements.txt
pycryptodome
pyqt5